pyproject.toml
[tool.poetry]
name = "heatmaps"
version = "0.1.0"
description = ""
authors = ["cs-gang"]

[tool.poetry.dependencies]
python = "^3.8"
sanic = "^21.3.4"
Jinja2 = "^2.11.3"
loguru = "^0.5.3"
motor = "^2.4.0"
docker-compose = "^1.29.1"
aiohttp = "^3.7.4"

[tool.poetry.dev-dependencies]
pre-commit = "^2.12.1"
black = "^20.8b1"
python-dotenv = "^0.17.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
